+++ package.json
@@ -10,6 +10,7 @@
     "classnames": "^2.3.2",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
+    "react-router-dom": "^6.5.0",
     "react-scripts": "5.0.1",
     "sass": "^1.56.2",
     "swiper": "^8.4.5",
