+++ src/components/Header/index.scss
@@ -1,139 +1,166 @@
-$burger-bar-height: 2px;
-$burger-bar-width: 50px;
+$burger-bar-height: 3px;
+$burger-bar-width: 45px;
 
 .header {
-    position: sticky;
-    top: 0;
+  position: sticky;
+  top: 0;
+  z-index: 9;
+  background: #ffffff;
+  box-shadow: 0px 4px 12px rgba(0, 0, 0, 0.1);
+  .container {
+    width: 1440px;
+    height: 80px;
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+    @media (max-width: 1440px) {
+      width: 100%;
+      padding: 0 20px;
+    }
+  }
+
+  &__block {
+    display: flex;
+    gap: 22px;
+    align-items: center;
+  }
+
+  &__burger {
+    display: flex;
     z-index: 9;
-    background-color: rgba(255, 255, 255, 0.75);
-    backdrop-filter: blur(10px);
-    box-shadow: 0 2px 4px 0 rgba(0, 0, 0, 0.1);
-    
-    .container {
-        display: flex;
-        justify-content: space-between;
-        align-items: center;
-        padding: 1rem 0;
+    flex-direction: column;
+    gap: 8px;
+    cursor: pointer;
+
+    & .burger__bar1,
+    & .burger__bar2,
+    & .burger__bar3 {
+      height: $burger-bar-height;
+      background-color: #494949;
+      transition: all 0.3s;
+      border-radius: 10px;
+    }
+    & .burger__bar1 {
+      width: $burger-bar-width;
+    }
+    & .burger__bar2 {
+      width: $burger-bar-width * 0.6;
+      margin: 0 auto;
+    }
+    & .burger__bar3 {
+      width: $burger-bar-width;
     }
 
-    &__burger {
-        display: flex;
-        z-index: 9;
-        flex-direction: column;
-        gap: 6px;
-        cursor: pointer;
-        padding: 10px;
-        
-        & .burger__bar1, & .burger__bar2, & .burger__bar3 {
-            background-color: #25262A;
-            transition: 0.2s;
-        } & .burger__bar1 {
-            width: $burger-bar-width;
-            height: $burger-bar-height + 1;
-        } & .burger__bar2 {
-            width: $burger-bar-width * 0.7;
-            height: $burger-bar-height;
-        } & .burger__bar3 {
-            width: $burger-bar-width;
-            height: $burger-bar-height + 1;
-        }
-        
-        
-        &.open .burger__bar1 {
-            transform: translate(-3px, 9px) rotate(-45deg);
-        }
-        
-        &.open .burger__bar2 {
-            opacity: 0;
-        }
-        
-        &.open .burger__bar3 {
-            transform: translate(-2px, -8px) rotate(45deg);
-            height: $burger-bar-height + 1;
-        }
+    &.open .burger__bar1 {
+      transform: translate(-5px, 10px) rotate(-45deg);
     }
 
-    &__title {
-        font-size: 3rem;
-        font-weight: 600;
+    &.open .burger__bar2 {
+      opacity: 0;
+    }
+
+    &.open .burger__bar3 {
+      transform: translate(-5px, -12px) rotate(45deg);
+    }
+  }
+
+  &__title h1 {
+    font-style: normal;
+    font-weight: 600;
+    font-size: 25px;
+    line-height: 41px;
+    letter-spacing: 0.02em;
+    color: #22242a;
+    transition: all 0.3s;
+    &:hover {
+      opacity: 0.8;
+    }
+    a {
+      color: inherit;
+    }
+  }
+
+  &__tools {
+    display: flex;
+    align-items: center;
+    gap: 12px;
+    padding: 6px 12px;
+    background: #ffff;
+    box-shadow: -2px -2px 10px rgba(0, 0, 0, 0.1), 2px 2px 10px rgba(0, 0, 0, 0.1);
+    border-radius: 10px;
+
+    [class^="tools__toggle-"] {
+      display: flex;
+      align-items: center;
+      width: 35px;
+      height: 35px;
+      justify-content: center;
+      gap: 4px;
+      padding: 4px;
+      background-color: #fff;
+      border-radius: 10px;
+
+      font-style: normal;
+      font-weight: 600;
+      font-size: 14px;
+      line-height: 16px;
+      letter-spacing: 0.01em;
+      color: #22242a;
+      transition: all 0.3s;
+
+      cursor: pointer;
+
+      &:hover {
+        opacity: 0.7;
+      }
+
+      // .toggle-theme__icon {
+      //     font-size: 1.5rem;
+      //     color: #000;
+      // }
+
+      // .toggle-theme__text {
+      //     font-size: 1.2rem;
+      //     font-weight: 600;
+      //     color: #000;
+      // }
+    }
+  }
+
+  &__menu {
+    position: fixed;
+    z-index: 8;
+    top: 70px;
+    transform: translateY(-100%);
+    opacity: 0;
+    visibility: hidden;
+    transition: 0.4s;
+    background-color: rgba(255, 255, 255, 0.97);
+    box-shadow: 0px 0px 15px 0px rgba(0, 0, 0, 0.2);
+    border-radius: 5px;
+    padding: 1rem;
+
+    &.open {
+      opacity: 1;
+      visibility: visible;
+      transform: translateY(calc(0% + 5px));
+    }
+
+    .menu__list {
+      display: flex;
+      flex-direction: column;
+      gap: 1rem;
+
+      .menu__item a {
+        font-size: 1.5rem;
+        font-weight: 500;
+        transition: 0.2s;
         color: #000;
-        transition: .2s;
-        
+
         &:hover {
-            color: #5C369A;
+          color: #5c369a;
         }
-        a {
-            color: inherit;
-        }
+      }
     }
-
-    &__tools {
-        display: flex;
-        align-items: center;
-        gap: 8px;
-        padding: 4px;
-        background-color: #000;
-        border-radius: 40px;
-
-        [class^="tools__toggle-"] {
-            display: flex;
-            align-items: center;
-            width: 35px;
-            height: 35px;
-            justify-content: center;
-            gap: 4px;
-            padding: 4px;
-            background-color: #fff;
-            border-radius: 50%;
-            cursor: pointer;
-
-            // .toggle-theme__icon {
-            //     font-size: 1.5rem;
-            //     color: #000;
-            // }
-
-            // .toggle-theme__text {
-            //     font-size: 1.2rem;
-            //     font-weight: 600;
-            //     color: #000;
-            // }
-        }
-    }
-
-    &__menu {
-        position: fixed;
-        z-index: 8;
-        top: 70px;
-        transform: translateY(-100%);
-        opacity: 0;
-        visibility: hidden;
-        transition: .4s;
-        background-color: rgba(255, 255, 255, 0.97);
-        box-shadow: 0px 0px 15px 0px rgba(0, 0, 0, 0.2);
-        border-radius: 5px;
-        padding: 1rem;
-        
-        &.open {
-            opacity: 1;
-            visibility: visible;
-            transform: translateY(calc(0% + 5px));
-        } 
-
-        .menu__list {
-            display: flex;
-            flex-direction: column;
-            gap: 1rem;
-
-            .menu__item a {
-                font-size: 1.5rem;
-                font-weight: 500;
-                transition: .2s;
-                color: #000;
-
-                &:hover {
-                    color: #5C369A;
-                }
-            }
-        }
-    }
+  }
 }
