+++ src/pages/Page404/Page404.scss
@@ -1,0 +1,51 @@
+.not-found {
+  position: relative;
+  width: 100%;
+  height: auto;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  img {
+    margin-top: 20px;
+    width: 750px;
+    height: auto;
+    @media (max-width: 790px) {
+      width: 100%;
+    }
+  }
+  &__title,
+  &__subtitle {
+    font-style: normal;
+    font-weight: 600;
+    font-size: 40px;
+    line-height: 54px;
+    letter-spacing: 0.02em;
+    text-align: center;
+    color: #494949;
+  }
+
+  &__title {
+    @media (max-width: 640px) {
+      line-height: 35px;
+      font-size: 25px;
+    }
+  }
+  &__subtitle {
+    margin-top: 12px;
+    font-weight: 400;
+    font-size: 20px;
+    line-height: 27px;
+    color: #636363;
+    @media (max-width: 640px) {
+      margin-top: 5px;
+      font-size: 16px;
+    }
+    a {
+      color: #5d25b9;
+      transition: all 0.3s;
+    }
+    a:hover {
+      opacity: 0.7;
+    }
+  }
+}
